+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Added `cmem_base_uri`, `dp_api_endpoint`, and `di_api_endpoint` to `SystemContext`.
 - Added `refresh_token` to `UserContext`.
 
+### Changed
+
+- Updated template 
+
 ## [4.15.0] 2025-10-15
 
 ### Changed